requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
pymysql==1.1.1
jinja2==3.1.4
python-multipart==0.0.9
bcrypt==4.1.3
itsdangerous==2.2.0
openpyxl==3.1.2
python-dotenv==1.0.1
Pillow==10.3.0
starlette==0.37.2
opencv-python-headless>=4.9.0
